Density of States (DOS)

Functions relevant to the plotting DOS distributions.

PlotDFT.import_DOS_VASP — Function

import_DOS_VASP(directory::AbstractString=="") -> DOSinfo

Imports information for plotting DOS from the VASP files DOSCAR, POSCAR, and OUTCAR and returns it as a DOSinfo object.

PlotDFT.import_DOS_lobster — Function

import_DOS_lobster(directory::AbstractString=="") -> DOSinfo

Imports information for plotting DOS from DOSCAR.lobster and POSCAR

PlotDFT.DOSinfo — Type

DOSinfo(
    tdos::DensityOfStates,
    pdos::Vector{ProjectedDensityOfStates},
    fermi::Real,
    alphabeta::Real,
    pos::PeriodicAtomList{3},
    )

Organizes information relevant to the plotting of the electronic density of states. See Electrum.jl's documentation for information about DensityOfStates, ProjectedDensityOfStates, and PeriodicAtomList.

PlotDFT.plot_DOS — Function

plot_DOS(
    dosinfo::DOSinfo;
    emin::Real=0,
    emax::Real=0,
    xmax::Real=0,
    eaxis::String="relative"
)
-> PlotlyJS.SyncPlot

Returns a plot of the total density of states.

PlotDFT.plot_pDOS — Function

plot_pDOS(
    plot::PlotlyJS.SyncPlot,
    dosinfo::DOSinfo;
    atom::Int,
    pdos::String,
    color::String="black")
-> PlotlyJS.SyncPlot

Adds a filled projected density of states to a given plot.

PlotDFT.dos_layout — Function

dos_layout(emin::Real, emax::Real, xmax::Real)

Returns a PlotlyJS layout object with default settings and ranges specified by emin/emax/xmax.

PlotDFT.num_electrons_at_energy — Function

num_electrons_at_energy(tdos::DensityOfStates, energy::Real) -> Float64

Returns the electron count of the unit cell at a specified energy using the total DOS.

PlotDFT.energy_at_electron_ct — Function

energy_at_electron_ct(dosinfo::DOSinfo, electron_ct::Real) -> Float64

Returns the energy of corresponding electron count using the total DOS.

energy_at_electron_ct(
    dosinfo::DOSinfo,
    electron_ct::Real,
    plot::PlotlyJS.SyncPlot;
    color::String="red"
)
-> PlotlyJS.SyncPlot

Adds a horizontal dotted line to a plot corresponding to the energy at a specified electron count.